refactor(shop): 로또 상점, 로또 일급 컬렉션 객체 패키지 분리
상점에서만 인증 된 Lottos를 생성 가능하도록 접근 제어
