+++ src/main/java/lotto/Application.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.shop.LottoShop;
+import lotto.shop.Lottos;
+
 public class Application {
 
 	public static void main(String[] args) {
