+++ src/test/java/lotto/GameResultTest.java
@@ -17,6 +17,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import lotto.shop.Lottos;
+
 public class GameResultTest {
 
 	Lottos lottos;
